project/requirements.txt
asgiref==3.8.1
azure-core==1.30.2
azure-storage-blob==12.22.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.0.9
django-storages==1.14.4
gunicorn==23.0.0
idna==3.8
isodate==0.6.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
python-decouple==3.8
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
